Create Common.Auth
