+++ Common/Auth/Extensions/ServiceExtensions.cs
@@ -1,43 +1,13 @@
-﻿using Identity.Data;
-using Identity.Entities;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Tokens;
 using System.Text;
 
-namespace Identity.Extensions
+namespace Common.Auth.Extensions
 {
-    public static class IdentityExtensions
+    public static class ServiceExtensions
     {
-        public static IServiceCollection ConfigurePersistence(this IServiceCollection services, IConfiguration configuration)
-        {
-            services.AddDbContext<IdentityContext>(options =>
-            {
-                options.UseSqlServer(configuration.GetConnectionString("IdentityConnectionString"));
-            });
-            services.AddScoped<IIdentityRepository, IdentityRepository>();
-
-            return services;
-        }
-
-        public static IServiceCollection ConfigureIdentity(this IServiceCollection services)
-        {
-            services.AddIdentity<User, IdentityRole>(options =>
-            {
-                options.Password.RequireLowercase = false;
-                options.Password.RequireUppercase = false;
-                options.Password.RequireDigit = true;
-                options.Password.RequireNonAlphanumeric = false;
-                options.Password.RequiredLength = 8;
-                options.User.RequireUniqueEmail = true;
-            })
-                .AddEntityFrameworkStores<IdentityContext>()
-                .AddDefaultTokenProviders();
-
-            return services;
-        }
-
         public static IServiceCollection ConfigureJWT(this IServiceCollection services, IConfiguration configuration)
         {
             var jwtSettings = configuration.GetSection("JWT");
