+++ Security/Identity/Extensions/ServiceExtensions.cs
@@ -1,0 +1,38 @@
+﻿using Identity.Data;
+using Identity.Entities;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+
+namespace Identity.Extensions
+{
+    public static class ServiceExtensions
+    {
+        public static IServiceCollection ConfigurePersistence(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.AddDbContext<IdentityContext>(options =>
+            {
+                options.UseSqlServer(configuration.GetConnectionString("IdentityConnectionString"));
+            });
+            services.AddScoped<IIdentityRepository, IdentityRepository>();
+
+            return services;
+        }
+
+        public static IServiceCollection ConfigureIdentity(this IServiceCollection services)
+        {
+            services.AddIdentity<User, IdentityRole>(options =>
+            {
+                options.Password.RequireLowercase = false;
+                options.Password.RequireUppercase = false;
+                options.Password.RequireDigit = true;
+                options.Password.RequireNonAlphanumeric = false;
+                options.Password.RequiredLength = 8;
+                options.User.RequireUniqueEmail = true;
+            })
+                .AddEntityFrameworkStores<IdentityContext>()
+                .AddDefaultTokenProviders();
+
+            return services;
+        }
+    }
+}
