+++ Security/Identity/Identity.csproj
@@ -23,4 +23,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\Common\Auth\Auth.csproj" />
+  </ItemGroup>
+
 </Project>
